チケット #35744 VS2015の警告除去
データ型の違いによる警告を除去した。


git-svn-id: https://svn.osdn.net/svnroot/ttssh2/branches/vs2015_warn@6240 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ ttssh2/ttxssh/ed25519_bcrypt_pbkdf.c
@@ -74,10 +74,10 @@ bcrypt_hash(u_int8_t *sha2pass, u_int8_t *sha2salt, u_int8_t *out)
 
 	/* key expansion */
 	Blowfish_initstate(&state);
-	Blowfish_expandstate(&state, sha2salt, shalen, sha2pass, shalen);
+	Blowfish_expandstate(&state, sha2salt, (u_int16_t)shalen, sha2pass, (u_int16_t)shalen);
 	for (i = 0; i < 64; i++) {
-		Blowfish_expand0state(&state, sha2salt, shalen);
-		Blowfish_expand0state(&state, sha2pass, shalen);
+		Blowfish_expand0state(&state, sha2salt, (u_int16_t)shalen);
+		Blowfish_expand0state(&state, sha2pass, (u_int16_t)shalen);
 	}
 
 	/* encryption */
